feat!: do not use subscriptions for rpc methods that send only one notification
Previously all rpc methods were subscriptions that returned a
`SubscriptionId`, and later reported on the progress of the task by
sending one of more notifications. Some methods like `list`, `rename`,
`create`, `delete` only ever sent one notification when the task was
completed. Such methods are now normal rpc methods that return a
response with the result of the task instead of a `SubscriptionId`.

Previously all methods were made subscriptions to make it easy for the
tasks related to these methods to be queued. But now I think that such
queuing of tasks should be handled by the client instead of the server,
therefore the module `task` that would have been responsible for such
task management is also removed.

Last but not least methods `create` and `delete` are also removed,
instead now there are different methods for creating and deleting files
and directories.

- method `create_file` to create new files
- method `create_dir` to create new directories
- method `delete_file` to delete files
- method `delete_dir` to delete empty directories